refactor: move credential related models to separate module
