+++ crates/cloud-wallet-openid4vc/src/oid4vci/credential/mod.rs
@@ -1,0 +1,7 @@
+pub mod formats;
+pub mod offer;
+mod request;
+mod response;
+
+pub use request::*;
+pub use response::*;
