+++ crates/cloud-wallet-openid4vc/src/oid4vci/mod.rs
@@ -1,12 +1,6 @@
-//! OpenID4VCI protocol data models.
-
 pub mod authorization;
 pub mod client;
 pub mod credential;
-pub mod credential_formats;
-pub mod credential_offer;
-pub mod credential_request;
-pub mod credential_response;
 pub mod error;
 pub mod metadata;
 pub mod notification;
